hacks to enable stable diffusion on Pixel backend
